autotest: move script helper functions from rover to common
